restructure tests/ to mirror the src/ package layout
Move each test file to the path matching the module it covers, and split the
files that mixed several modules, so tests/ mirrors src/prpl_kinematics/
(geometry/, tree/, loading/, planning/, ik/, robots/, visualization/). Pure
move/rename/split -- no test behavior changes; conftest.py stays at the root.

- test_planning.py -> planning/{test_joint_space,test_se2_space,test_birrt}.py
- test_ik.py       -> ik/{test_numerical,test_follow}.py
- test_visualization.py -> test_meshes.py + geometry/test_shapes.py
                          + visualization/{test_pybullet_renderer,test_blender_renderer,test_interface}.py
- simple relocations for the rest (collision/manipulation stay at the root,
  matching their root modules).

Closes #492.
