+++ prpl-kinematics/tests/ik/test_follow.py
@@ -1,4 +1,4 @@
-"""Unit tests for numerical inverse kinematics and end-effector following."""
+"""Unit tests for end-effector path following."""
 
 import os
 
@@ -7,11 +7,7 @@ import pytest
 from scipy.spatial.transform import Rotation
 from spatialmath import SE3
 
-from prpl_kinematics.ik import (
-    InverseKinematics,
-    NumericalIK,
-    follow_end_effector_path,
-)
+from prpl_kinematics.ik import NumericalIK, follow_end_effector_path
 from prpl_kinematics.loading import load_urdf
 from prpl_kinematics.planning import JointSpace
 from prpl_kinematics.utils import get_assets_path
@@ -44,36 +40,6 @@ def _pose_error(tree, ee, config, target) -> tuple[float, float]:
     position = float(np.linalg.norm(np.asarray(target.t) - np.asarray(reached.t)))
     rotation = Rotation.from_matrix(np.asarray(target.R) @ np.asarray(reached.R).T)
     return position, float(np.linalg.norm(rotation.as_rotvec()))
-
-
-def test_numerical_ik_conforms_to_interface():
-    """NumericalIK satisfies the InverseKinematics protocol."""
-    tree, space = _panda()
-    assert isinstance(NumericalIK(tree, space, "panda_hand"), InverseKinematics)
-
-
-def test_solve_reaches_nearby_pose():
-    """IK recovers a configuration reaching a pose from a nearby seed."""
-    tree, space = _panda()
-    ik = NumericalIK(tree, space, "panda_hand")
-    truth = _comfortable()
-    target = tree.forward_kinematics("panda_hand", truth)
-    seed = {name: [value[0] + 0.3] for name, value in truth.items()}
-    solution = ik.solve(target, seed)
-    assert solution is not None
-    position_error, orientation_error = _pose_error(
-        tree, "panda_hand", solution, target
-    )
-    assert position_error < 1e-3 and orientation_error < 1e-2
-    vector = space.to_vector(solution)
-    assert np.allclose(vector, space.clamp(vector))
-
-
-def test_solve_returns_none_when_unreachable():
-    """A target far outside the workspace yields no solution."""
-    tree, space = _panda()
-    ik = NumericalIK(tree, space, "panda_hand")
-    assert ik.solve(SE3(5.0, 5.0, 5.0), _comfortable()) is None
 
 
 def test_follow_end_effector_path_is_smooth():
