+++ prpl-kinematics/tests/robots/test_panda.py
